Move ValidationRegexp to common
