--- cab-app-passenger-service/src/main/java/by/arvisit/cabapp/passengerservice/util/ValidationRegexp.java
@@ -1,9 +1,0 @@
-package by.arvisit.cabapp.passengerservice.util;
-
-public final class ValidationRegexp {
-
-    public static final String CARD_NUMBER = "[0-9]{16}";
-    
-    private ValidationRegexp() {}
-
-}
