+++ cab-app-payment-service/src/main/java/by/arvisit/cabapp/paymentservice/dto/PassengerPaymentRequestDto.java
@@ -1,8 +1,8 @@
 package by.arvisit.cabapp.paymentservice.dto;
 
+import static by.arvisit.cabapp.common.util.ValidationRegexp.CARD_NUMBER;
 import static by.arvisit.cabapp.paymentservice.persistence.model.PaymentMethodEnum.BANK_CARD;
 import static by.arvisit.cabapp.paymentservice.persistence.model.PaymentMethodEnum.CASH;
-import static by.arvisit.cabapp.paymentservice.util.ValidationRegexp.CARD_NUMBER;
 
 import java.math.BigDecimal;
