--- cab-app-payment-service/src/main/java/by/arvisit/cabapp/paymentservice/util/ValidationRegexp.java
@@ -1,9 +1,0 @@
-package by.arvisit.cabapp.paymentservice.util;
-
-public final class ValidationRegexp {
-
-    public static final String CARD_NUMBER = "[0-9]{16}";
-
-    private ValidationRegexp() {
-    }
-}
